+++ src/main/java/com/besson/arknights/block/custom/columbian/PottedLongLeafPlant.java
@@ -1,0 +1,21 @@
+package com.besson.arknights.block.custom.columbian;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class PottedLongLeafPlant extends FurnitureHorizontalFacingBlock {
+    public PottedLongLeafPlant(Settings settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0,0,0,16,24,16);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}